bcachefs: distribute ios to caches better
Uses a random distrobution weighted by amount of free buckets
to select which cache.

Cache selection will wait, like bch_bucket_alloc, if wait paramenter
is passed in. It will fail likewise as well.

Only the first ptr is garunteed to go to the desired cache, the rest
may be allocated to any tier.

Signed-off-by: Nicholas Swenson <nks@daterainc.com>
Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>