bcachefs: Replicate writes to multiple cache devs
This mostly invoves a new function to handle the submission of bios
to multiple ptrs in a key. It simply clones the original bio, and submits
it to all the devs a key points to.

The moving_gc write path needed changes to handle moving a single key
while preserving ptrs to other devices, and to handle the possibility of
moving multiple keys. To do this, I detached the allocation paths of
foreground and gc wrties. Perhaps later they can be realigned but for
now this is simplest.

Signed-off-by: Nicholas Swenson <nks@daterainc.com>
Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>