bcachefs: put a closure inside btree_op
This closure is used to wait on btree node allocation. The code around
the bch_data_insert_keys() function has been redone to use this closure.

Also, the cache_lookup() function in request.c is now called by s->op.cl
and not s->iop.cl. This continues the effort to decouple the read path
from the data_insert_op, which should only be used for writes.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>