bcachefs: allocator triggers tiering and moving GC
Previously, the allocator thread would wake up btree GC when it could
not find sufficient buckets to invalidate. If we were in a situation
where only tiering or moving GC could make space, this behavior was
counter-productive -- btree GC would be woken up over and over again,
and since it currently holds a write lock on the root, it would
throttle writes issued by tiering and moving GC, while itself failing
to make any progress.

Instead, we now wake up tiering if we're out of space on a lower tier,
and wake up moving GC if we're out of space on the highest tier. To
prevent CPU spinning, the allocator then waits for the next btree GC
to run. Tiering and moving GC kick off a btree GC either if they
themselves run out of space, or when they complete a pass.

Finally, wake up moving GC every 30 seconds to try to do some work,
otherwise we would never do any moving GC on any device not in the
highest tier.

We still need to tune the PD controllers for tiering and moving GC,
but at least btree GC no longer starts spinning when we're out of
space.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>