bcachefs: don't count stale pointers towards the replicas_needed total
This is a subtle corner case that can only come up during initial GC,
which occurs before journal replay.

More specifically, the initial btree on disk might contain a key with
two pointers [X, Y] where Y is on tier 1, and in between the time when
the btree node was last written to disk and when we crashed, Y was
copied to a new location.

When we replay the journal, we will insert a key with pointers [X, Y']
where Y' is the new copy of Y. This replaces the old key, [X, Y].
Depending on if the btree node has been compacted or not, we might see
the old key as having a single pointer [X] instead, since compaction
drops stale pointers.

The sector subtraction thus might see that X is the only copy, and treat
it as dirty. This means when we mark [X, Y], we must also treat X as
dirty since Y is already stale, otherwise the subtraction hits an
assert checking for a non-negative dirty sector count.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>