bcachefs: fix sector counting bug
This is a regression from "journal key inserts when the key is a
split of the insert key". If we have to bail out of the extent
fixup because there is no room in the journal reservation, we
have to subtract the sectors we haven't inserted yet, otherwise
they will be double-counted later.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>