bcachefs: fixing EXTENT_CACHED usage
Recently, KEY_CACHED was dropped in favor of EXTENT_CACHED, which uses
the most significant bit of the first pointer.

When doing a write, we can't set the key to cached in bch_write_op_init()
because it may not have any pointers, instead do it after allocating
sectors.

Also, we have to save and restore cached status in a few places where
pointers are re-ordered or replaced:

- When sorting pointers in bch_extent_normalize()

- When replacing a pointer in moving GC

This fixes a regression from "New bkey format".

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>