bcachefs: fix tiering space check, remove special btree node reserves
Tiering needs to block if tier 1 has fewer than 10% buckets free, since
in this case we have to wait for moving GC to catch up otherwise we
can deadlock.

After recent changes to add external iterators, the btree_alloc_reserve
field of bch_write_op was no longer used. Furthermore, only tiering
was setting it, and moving GC was not. Finally, the comment explaining
RESERVE_MOVING_GC_BTREE was wrong, because the size of the FIFO was the
same as the BTREE_ID_EXTENTS reserve, and not the same as free_inc.

The original motivation for these reserves was an incomplete fix for
-ENOSPC issues. Now that this is handled better, just nuke all this dead
code.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>