bcachefs: btree_split() fix
When updating interior nodes, btree_split() has to do the insert before
splitting and making the two new nodes visible - because the insert has to be
atomic, and post split it might not be.

But before the split, there might not be enough room, and with coalescing it's
not really sane to guarantee there will always be enough room - this is the bug
that this patch fixes.

But it is ok to do the insert post split, _provided_ the insert happens before
the two new nodes are visible (via updating the parent node).

Also, make sure insert_at() doesn't return 0 when it didn't finish.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>