bcachefs: Always preserving ordering after unclean shutdown
It's a very desirable property that order of index updates is always preserved
after a crash: i.e. if before the crash there were updates 1, 2, and 3, after
journal replay if 2 didn't make it (which is fine if the journal wasn't flushed)
3 shouldn't have made it either.

This is needed for bcachefs - we want to ensure that we never have dirents
pointing to nonexistent inodes after a crash, so we need to strictly order the
dirent and inode i_nlink updates.

To accomplish this, recall that the journal already implements a total order of
all index updates: we need to ensure that if an index update's journal entry
didn't make it after a crash, it doesn't show up - which currently could happen
if the btree node was flushed before the journal entry.

We simply have to record, in each bset (btree node log entry), the max sequence
number of any journal entry its updates correspond to. Then, during recovery, we
ignore bsets with journal_seq > the sequence number of the most recent journal
entry we found.

This won't cause us to lose needed btree node updates because before the bset
was written, all the journal entries it had index updates for were considered
dirty; those journal entries will still be considered dirty until a journal
entry is written _after_ that bset's write has completed.

The complicated part is that we then need to blacklist those journal sequence
numbers - we can't start writing journal entries with the next sequence number
where we left off, or the next time we crash and recover we're going to now use
bsets that should still be blacklisted. That's where the bulk of the machinery
in this patch comes in.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>