bcachefs: Refcount outstanding writes to safely go read only
Based on Bill Rozas's patch "bcache: Add an interlock for turning off cache
sets" but simplified.

We shouldn't need a ref like this for outstanding reads, because to do a read
you have to have the cache set open (holding a ref on the cache set), which will
prevent it from going away. But we do need this for writes, in order to safely
go RO - we don't want to e.g. shutdown gc/copygc/tiering before stopping
foreground writes because then we have a potential deadlock, and in particular
we were hitting a bug where a foreground write went to do a btree split, but the
allocators for all the devices shut down in between the split checking the btree
reserve and doing the allocation.

We should be able to cope with that without crashing without this patch, but
really we don't want to attempt the split at all in that case - so we still want
this patch.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>