bcachefs: Kill bch_bset_insert_with_hint(), simplify insert path
instead of doing another search, add bch_btree_node_insert_pos() to get the
insert position out of the iterator.

Also, no real need for insert_fixup() to be a method - btree_insert_key() has a
conditional for if (is_extents) right there - it can just call the right fixup
itself, then we can delete some not-really-needed helpers. Kind of moving to a
model where insert_fixup() is the toplevel of the insert, and bch_bset_insert()
is the only insert method bset.c has to expose, which is nifty.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>