bcachefs: Coalescing fix
This fixes a regression from "linked iterators" - that patch dropped the code in
bch_coalesce_nodes() to set iter->pos to the first key being inserted and
restore it afterwards - instead it was only reiniting the node iterator. That
was incorrect, because inserting might compact/sort bsets in the node being
inserted into and have to reinit the node iterator again, and that would use the
btree iter position.

But this is potentially a problem for btree_split() as well - when splitting
nodes the key for the new node also comes before the current iterator position,
we were just relying on the fact that there wouldn't be any keys between the
iterator's current position and where the key for the new node should go.

But this is potentially wrong, if we correctly take into account deleted keys -
for non extents, deleted keys come before non deleted keys, so if there is a
duplicate key at the current iterator position the node iterator is going to
point after it - and the key for the new node would be inserted at the wrong
position.

This hasn't been happening due to pure dumb luck - since for interior nodes we
currently write node after every insert, we won't ever insert the key for the
new node into the same bset as the deleted key it would have been incorrectly
inserted after.

Since this is only an issue for updates to interior nodes, rather than play
games with the iterator position just handle this in
bch_btree_insert_keys_interior() by manually backing up the node iterator as
needed - rewinding a btree node iterator is always safe because
bch_btree_iter_traverse() will advance it as needed (in case iter->pos changed).