bcachefs: Centralize bch_keybuf_check_overlapping()
bch_keybuf_check_overlapping() is used as an optimization to if possible keep
copygc from moving data around that we're about to overwrite (note that it's not
an optimization when used on the writeback keybuf! that's critical for cache
coherency).

Anyways, it goes with writing data to the cache, so stick it there.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>