bcachefs: Add a workqueue for btree insertions
The next patch is going to make struct open_bucket something with a refcount
that gets allocated, and the refs will be dropped after we do the index update
(in bch_data_insert_keys()); we can't drop the ref (i.e. freeing it) from the
same workqueue as we do the allocation - that would block frees - so add a new
workqueue for bch_data_insert_keys().

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>