bcachefs: Wait on journal write with correct closure
When we need to do a sync write (i.e. wait on the journal write to hit disk), we
do that by passing a closure to the journalling code to go on a waitlist.

bch_btree_insert_node() was using the wrong closure - a closure can only be on
one waitlist at a time, and the closure in struct btree_op can be on other
waitlists. Make the closure an explicit argument so the caller can pass one
that's safe to use.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>