bcachefs: bch_btree_iter_next_overlapping() -> bch_btree_iter_peek_overlapping()
This is prep work for doing multiple insertions without reiniting the btree
iterator.

We need to not advance past the current key unless we actually do overlap with
the key after it - because the next key we insert might overlap with it too.
Hence, the fix_overlapping() functions have to be reworked to use peek.

While we're at it, make some improvements to bkey_cmpxchg() and
handle_existing_key_newer().

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>