bcachefs: Fix bch_generic_insert_fixup()
iter_peek() will advance the iterator past where we want it (until we see a non
deleted key), potentially screwing up the insert. this hasn't technically been a
problem before, but it was a bug waiting to happen, and with how the next patch
will use the iterator it needs to be fixed.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>