bcachefs: Fix device add
Adding a device to an existing filesystem now works again - fixed a
number of bugs related to multiple devices (when mounting a filesystem
with multiple devices, we weren't actually checking that all the devices
were part of the same filesystem!).

Also did some refactoring to work towards being able to run in degraded
mode.