bcachefs: Take bucket_gc_gen() into account in allocator
The allocator used to take bucket_gc_gen() into account when deciding
which buckets to use - this helps to avoid reusing the same buckets
repeatedly, as when that happens we have to trigger a mark and sweep gc
to avoid generation number wraparound.

That was dropped in e872b56a31 "bcachefs: Kill bucket_mark.copyg", but
really shouldn't have been - readd that.